# Conflicts:
#	ewm-service/service/src/main/java/ru/practicum/mapper/EventMapper.java
#	ewm-service/service/src/main/java/ru/practicum/model/dto/EventFullDTO.java

+++ ewm-service/service/src/main/java/ru/practicum/service/UserService.java
@@ -9,6 +9,6 @@ public interface UserService {
 
   void delete(Long userId);
 
-  List<UserDTO> getUsers(List<Integer> ids, Integer from, Integer size);
+  List<UserDTO> getUsers(List<Long> ids, Integer from, Integer size);
 
 }
